refactored to use new architecture scheme
