+++ client/src/Components/App/App.component.tsx
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import logo from "./logo.svg";
 import "./App.css";
 
-class App extends Component {
+export class App extends Component {
   public render() {
     return (
       <div className="App">
@@ -24,5 +24,3 @@ class App extends Component {
     );
   }
 }
-
-export default App;
