+++ client/src/Components/App/index.ts
@@ -1,0 +1,1 @@
+export * from "./App.component";
