+++ client/src/index.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-import App from "./Components/App";
+import { App } from "./Components";
 import "./index.css";
 import * as serviceWorker from "./serviceWorker";
